chore: fix eip unit test issue
